src/serpapi-agent/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
langgraph>=0.2.0

# Natural language parsing with Claude (same style as brave-search-agent)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0

# Browser automation (HEADLESS / headed)
playwright>=1.49.0
